july-2021-leetcode-challenge/src/main/java/org/phoenix/leetcode/challenges/Problem26_ConvertSortedArrayToBinarySearchTree.java
package org.phoenix.leetcode.challenges;

/**
 * Given an integer array nums where the elements are sorted in ascending order, convert it to a height-balanced binary search tree.
 * <p>
 * A height-balanced binary tree is a binary tree in which the depth of the two subtrees of every node never differs by more than one.
 * Constraints:
 * <p>
 * 1 <= nums.length <= 104
 * -104 <= nums[i] <= 104
 * nums is sorted in a strictly increasing order.
 */

public class Problem26_ConvertSortedArrayToBinarySearchTree {
    public TreeNode sortedArrayToBST(int[] nums) {
        if (nums.length == 0) {
            return null;
        }
        return helper(nums, 0, nums.length - 1);
    }

    public TreeNode helper(int[] nums, int low, int high) {
        // Base condition
        if (low > high) {
            return null;
        }
        // Middle element
        int mid = (low + high) / 2;
        // Creating root element
        TreeNode node = new TreeNode(nums[mid]);
        // Connecting left and right subtree
        node.left = helper(nums, low, mid - 1);
        node.right = helper(nums, mid + 1, high);
        return node;
    }


    // Definition for a binary tree node.
    public static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;

        TreeNode(int val) {
            this.val = val;
        }
    }
}
